requirements.txt
streamlit
psycopg2-binary
bcrypt
python-dotenv
pandas